backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sashecka-backend"
version = "0.1.0"
description = "Minimal FastAPI backend with auth and Swagger"
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.0",
  "pydantic>=2.7.0",
  "email-validator>=2.0.0",
  "PyJWT>=2.8.0",
  "pwdlib[argon2]>=0.2.0",
  "python-multipart>=0.0.9",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.uv]
package = true
